IRG/irg_submission/yolodeployment/requirements.txt
torch==1.9.0
Pillow
fastapi
uvicorn==0.11.7
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torchvision>=0.8.1
tqdm>=4.41.0
seaborn>=0.11.0
pandas
tensorboard>=2.4.1
